backend/requirements.txt
# SRP AI HRMS Backend - Python Dependencies
# Python 3.12+

# ── Framework ──────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ── Database ───────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
greenlet==3.1.1

# ── Validation / Settings ──────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# ── Auth & Security ────────────────────────────────────────────────────────
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# ── Redis (caching / token blacklist) ─────────────────────────────────────
redis==5.2.1
hiredis==3.1.0

# ── Background Tasks ───────────────────────────────────────────────────────
celery==5.4.0
flower==2.0.1

# ── HTTP Client ────────────────────────────────────────────────────────────
httpx==0.28.1
aiohttp==3.11.11

# ── File Storage ──────────────────────────────────────────────────────────
boto3==1.35.86
botocore==1.35.86
aiofiles==24.1.0

# ── Logging ────────────────────────────────────────────────────────────────
structlog==24.4.0
python-json-logger==3.2.0

# ── Utilities ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1
python-slugify==8.0.4
Pillow==11.1.0
pytz==2024.2
arrow==1.3.0
uuid6==2024.1.12

# ── Testing ────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
factory-boy==3.3.1
faker==33.1.0

# ── Dev / Docs ─────────────────────────────────────────────────────────────
setuptools==75.6.0
